requirements.txt
Flask==3.0.0
gunicorn==21.2.0
numpy==1.24.3
Pillow==10.0.0
scipy==1.11.2
reportlab==4.0.4